+++ CHANGES.md
@@ -5,6 +5,7 @@
 * A new setting allows to turn image smoothing of image layers on and off.
   By default it is off now, so that crisp image pixels are shown without any blurring. (#86)
 * The color bar legend uses 5 tick marks instead of the two.
+* Added simple scale indicator to map.
 * Now using OpenLayers 6.1.1.